requirements.txt
pysftp==0.2.9
numpy==1.18.2
netCDF4
cryptography==2.8
PyYAML==5.3.1
zipfile2==0.0.12
pillow
ipython
psycopg2
fiona
Shapely==1.8.0
xarray
rioxarray==0.8.0

